Automating the process for creating the JSON file needed for GitHub API
A script can be ran during installation to generate the ignored json
file necessary for the GitHub API to work properly.

A createGitHubJSON script has been created that will only run in a non-
production environment. A test has been created to ensure that the
dynamic data is generated properly.

The script takes the username as the first argument and the
authentication token if present as the second argument.

Instructions have been added to the README but the script can be invoked
by running the following command:

```bash
npm run create-github -u GitHubUsername -a GitHubAccessToken
```