Added a getRepoByName method to the GitHubAPI
The idea is to have a dynamic portfolio that defaults to the pinned
repositories. The getPinnedRepoNames, returns a list of the pinned
repo names and for each repo name we could call getRepoByName method
to fetch the data. This is a bit more efficient than fetching all
the repos and then filtering the pinned ones. Tests have were created
all all tests currently pass.