Working on Admin functionality. Currently a pre-authorized user can
create an account and login. Upon successful login or creation the user
is redirected to the admin dashboard, which is currently a blank page.
The API routes are protected like the others and requires the github
token to be passed in the header. But the admin dashboard is
authenticated by a JSON Web token only which is only valid for two hours