Slight housekeeping and minor refactoring.
Ensured the env references are consistent. NEXT_PUBLIC was added
to the local env file, but the test file doesn't need public variables
The password input validation incorrectly had less than or equal to,
this was changed to less than. The attache default JSON has been updated
to use two different files, one is the defaults while the other holds
information for the dynamic portfolios.