The ability to create attaches has been implemented
The form has a lot of steps and was split into several components
The dynamic resume portion still needs to be added but currently
it was more important to implement the ability to add and verify the
projects being added to the attache.

Only legitimate projects from the user's account can be added.
Once added that projects name is removed from the available list so
duplicates are not possible. Urls for the demo videos are also validated
they can be a legitimate URL or use the following: DEMO, GITHUB, NONE.
Once at least 6 but no more than 8 projects are added the user can
add in the identifying information for the attache and create it.

The next step should be to display the list of projects on the dashboard
so they can be edited or deleted as needed. This view will differ from
the Attache view in that it will be a list with editable options; while
the Attaches view is more of a portfolio style where the different
attaches could be viewed like the different projects of a portfolio.