Refactored the existing APIs to make the Admin and non-Admin APIs more
consistent. A lot of files were touched in this commit, as the
entire implementation was touched up a bit before writing more code
to fetch more data. Data fetching changes were also updated for the
URL for the AVATAR, this is handled by the Layout component now.
A slight issue with the existing attache creation was addressed and a
provider component was created to manage the state of the attaches on
the user's dashboard.