Updated ListState Provider to cache the retrieved list item data
The attaches are cached when created or when the dashboard does a cold
load. This should limit the number of calls to the server and improve
performance. The list item can now be styled and provide the necessary
information, all the logic for retrieving and caching the data has been
completed.