Individual Attache view id completed. Can access via list in dashboard
The individual attache can also be viewed via the projects/id
but using the id of the attache, I know that doesn't follow true REST
but someone viewing the profile isn't aware of attaches, the projects
endpoint already exists and for all intents and purposes, it is the same
thing. A lockout has been places so that if you are an end viewer and
you had accessed an attache via the projects/id, you can't access
the pinned repos by clicking on the projects nav link, it will point
to the dynamic projects/id endpoint instead (the whole point). This will
need to be refactored to use a version provider to truly lock it down
but for now, it works. The refactor would allow a /id endpoint to be
given out for dynamic access, makes a shorter URL.