Updated the data fetching logic for individual projects/repos.
Each fresh landing will pull fresh data from GitHub, if the data isn't
in the cache, the cache is volatile, but persists as long as the user
doesn't manually refresh the page, navigate away, close the tab, etc.
This keeps our data as fresh as possible while limiting the number of
requests we make to GitHub and our own API. We could always make the
rules more advanced but this is a good start.