Trying to correct an issue that presents itself only when deployed.
The Img for one of the repos is a large gif, and takes the largest
time to paint. This was addressed with the priority attribute for the
Next/Image component. And the warning message is gone locally.

The avatar was being fetched in two places, once for the layout
and was being sent as SSR props for the about page. This has been
abstracted to an avatar provider. It is loaded by the layout since it
appears on all pages.

Everything works and builds locally Hopefully this resolves the issue
with the Covid-Master project.