Still an issue with a pinned repo's screenshot when deployed.
The covid master project loads fine locally in dev mode, but maybe
the file is too large and the response times out. It errors in
production but I don't see relevant information as to why. This change
should just force the default image so it isn't blank but I will take
a new screenshot and upload it to the repo.