Updated the api/project/[name] endpoint to always fetch fresh data
from github, regardless if it is a dynamic request or not.
Dynamic Requests are used for all attache projects. While the default
landing page uses the static version of the endpoint, telling it
where to look for the DEMO URL. On the static version, if the data was
in the database, it wouldn't be updated unless it was added again as
a project in an attache. This could cause changes to not show, so it has
been changed for now. It could be improved upon in the future by
implementing a permanent cache, for now the implemented cache is
volatile.