Updating the README and adjusted the tests to account for the
change back to using just ENVS. A createEnv script has been created
and added to the package.json under env. This script will generated the
ENVs that need to be used for the project and takes the GitHub username
and auth token as arguments. Remaining ENVS are set in name only
where the user can set them to their own values.

The installation instructions have been updated to reflect these changes
and a test build was preformed locally to prior to this commit.