Added the Suspense component to the project card.
Wanted to use the full API with the resource convention but
that would require a major refactor, as we are using providers and
hooks to fetch and cache our data. Additionally, all data fetches are
dynamic and require information from the component before fetching.
Our current method ensures a single fetch and caches data so that
unless the page reloads, data is served from the cache. Suspense was
also added to our Avatar Image, but due to a typo, I maxed out the API
calls for the next hour, HA, and I am unsure if that will really improve
loading as it also sits behind a hook, but this could be easily
refactored if needed.