Updated date on license file.
The deployed site isn't updating the pinned repos as it should. They are
updating locally if I change a pinned Repo but not on the deployed site.

There is no reason I can see why this would be happening, we query a
list of repo names then look them up. This list of names is what isn't
updating.